requirements.txt
faiss-cpu>=1.8.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
langchain-text-splitters>=0.2.0
numpy>=1.26.0
pypdf>=4.0.0
requests>=2.32.0
sentence-transformers>=3.0.0
